Refactor problem module
